requirements.txt
awscli
aws-sam-cli
numpy==1.17.4
Pillow==6.2.1
pytest==5.1.3
